requirements.txt
Flask==1.0.2
connexion[swagger-ui]==2.0.1
SQLAlchemy==1.2.13
alembic==1.0.2
semver==2.8.1
pyyaml
babel==2.6.0
natsort==5.1.0
pycountry==18.12.8